Cleanup file structure of the NegativeCompileTestsRunner target
